+++ tests/unit/unit_test_client.py
@@ -125,51 +125,3 @@ def test_submit(*mocks):
     assert not get_mock.called
     POST_mock.assert_called_once_with('/evaluation/submission?etag=Fake eTag', None)
     assert submission_mock.called
-
-
-def test_with_retry():
-    retryParams = {"retries": 3, "wait": 0}
-    response = MagicMock()
-    function = MagicMock()
-    function.return_value = response
-    
-    # -- No failures -- 
-    response.status_code.__eq__.side_effect = lambda x: x == 250
-    syn._with_retry(function, **retryParams)
-    assert function.call_count == 1
-    
-    # -- Always fail -- 
-    response.status_code.__eq__.side_effect = lambda x: x == 503
-    syn._with_retry(function, **retryParams)
-    assert function.call_count == 1 + 4
-    
-    # -- Fail then succeed -- 
-    thirdTimes = [3, 2, 1]
-    def theCharm(x):
-        if x == 503:
-            count = thirdTimes.pop()
-            return count != 3
-        return x == 503
-    response.status_code.__eq__.side_effect = theCharm
-    syn._with_retry(function, **retryParams)
-    assert function.call_count == 1 + 4 + 3
-    
-    # -- Retry with an error message --
-    retryErrorMessages = ["Foo"]
-    retryParams["retry_errors"] = retryErrorMessages
-    response.status_code.__eq__.side_effect = lambda x: x == 500
-    response.headers.__contains__.reset_mock()
-    response.headers.__contains__.side_effect = lambda x: x == 'content-type'
-    response.headers.__getitem__.side_effect = lambda x: "application/json" if x == 'content-type' else None
-    response.json.return_value = {"reason": retryErrorMessages[0]}
-    syn._with_retry(function, **retryParams)
-    assert response.headers.__getitem__.called
-    assert response.headers.__getitem__.called
-    assert function.call_count == 1 + 4 + 3 + 4
-    
-    # -- Propagate an error up --
-    def foo(): raise SynapseError("Bar")
-    function.side_effect = foo
-    assert_raises(SynapseError, syn._with_retry, function, **retryParams)
-    assert function.call_count == 1 + 4 + 3 + 4 + 1
-
